% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_grtsdb.R
\name{apply_grtsdb}
\alias{apply_grtsdb}
\title{apply grtsdb}
\usage{
apply_grtsdb(perimeter, cellsize = 100, n = 20, export_path = ".", seed)
}
\arguments{
\item{perimeter}{a simple features (sf) object}

\item{cellsize}{an optional integer. The size of each cell. Either a single value or one value for each dimension. Passed onto extract_sample from GRTSdb. Default is 100.}

\item{n}{an optional integer. the sample size. Passed onto extract_sample from GRTSdb. Default is 20}

\item{export_path}{an optional character string pointing to the path where the GRTSdb.sqlite is created. Default is "."}

\item{seed}{a optional character. Allowing to rerun a previous use.}
}
\description{
Applies \code{grtsdb::extract_sample} from inbo/GRTSdb to a custom perimeter. This function installs
GRTSdb if it is missing from your machine.
}
\details{
A function to apply grtsdb to a custom perimeter

GRTSdb is automatically installed when missing from your system.
}
\examples{
\dontrun{
# Preparation
perimeter <- sf::st_as_sf(boswachterijen$boswachterijen_2024) \%>\%
  dplyr::filter(Regio == "Taxandria",
                Naam == "vacant 4")

# A new sample
sample <- apply_grtsdb(perimeter,
                       cellsize = 1000,
                       n = 20,
                       export_path = ".")

leaflet::leaflet() \%>\%
 leaflet::addTiles() \%>\%
 leaflet::addCircles(data = sample$samples,
                     color = "red") \%>\%
 leaflet::addPolylines(data = sample$grid,
                       color = "blue") \%>\%
 leaflet::addPolylines(data = perimeter,
                       color = "black")
# Reuse a old sample
seed <- sample$seed

sample <- apply_grtsdb(perimeter,
                       cellsize = 1000,
                       n = 20,
                       export_path = ".",
                       seed = seed)

 leaflet::leaflet() \%>\%
 leaflet::addTiles() \%>\%
 leaflet::addCircles(data = sample$samples,
                     color = "red") \%>\%
 leaflet::addPolylines(data = sample$grid,
                       color = "blue") \%>\%
 leaflet::addPolylines(data = perimeter,
                       color = "black")
}
}
\seealso{
Other spatial:
\code{\link[=CRS_extracter]{CRS_extracter()}},
\code{\link[=add_habitats]{add_habitats()}},
\code{\link[=calculate_polygon_centroid]{calculate_polygon_centroid()}},
\code{\link[=collect_osm_features]{collect_osm_features()}},
\code{\link[=dms_column_to_decimal]{dms_column_to_decimal()}},
\code{\link[=dms_to_decimal]{dms_to_decimal()}}
}
\author{
Sander Devisscher
}
\concept{spatial}
