% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLC18_BE.R
\docType{data}
\name{CLC18_BE}
\alias{CLC18_BE}
\title{CLC18_BE}
\format{
datapackage
\describe{
\item{ID}{Unieke identificatie polygonen}
\item{CODE_18}{CORINE landcover code 2018}
\item{LABEL_18}{CORINE landcover label 2018}
\item{LABEL_Grouped}{Lijst van media-urls van de geselecteerde deployments}
\item{geometry}{Geometrie kolom met polygonen}
\item{...}{Overige kolommen}
}
}
\usage{
CLC18_BE
}
\description{
Corine landcover of Belgium (2018)
}
\seealso{
Other dataset:
\code{\link{boswachterijen}},
\code{\link{drg_example}},
\code{\link{surveyMonkey_q6}}
}
\concept{dataset}
\keyword{datasets}
