% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_sqlite.R
\name{cleanup_sqlite}
\alias{cleanup_sqlite}
\title{cleanup sqlite}
\usage{
cleanup_sqlite(db = "grts.sqlite")
}
\arguments{
\item{db}{name of the temporary .sqlite db to be removed}
}
\description{
A helper script to cleanup after use of apply_gtrsdb.
}
