% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcompare.R
\name{colcompare}
\alias{colcompare}
\title{Columnname comparison}
\usage{
colcompare(x, y)
}
\arguments{
\item{x}{dataframe 1}

\item{y}{dataframe 2}
}
\value{
a list of columns present in x but not in y and a list of columns
present in y and not in x.
}
\description{
A simple function to list the difference in column names in 2 datasets.
}
\examples{
\dontrun{
# create example dataframes
super_sleepers <- data.frame(rating=1:4,
animal=c('koala', 'hedgehog', 'sloth', 'panda'),
country=c('Australia', 'Italy', 'Peru', 'China'),
avg_sleep_hours=c(21, 18, 17, 10))

super_actives <- data.frame(rating=1:4,
animal=c('kangeroo', 'wolf', 'jaguar', 'tiger'),
country=c('Australia', 'Italy', 'Peru', 'China'),
avg_active_hours=c(16, 15, 8, 10))

colcompare(super_sleepers, super_actives)
}

}
\seealso{
Other dataframe_comparison:
\code{\link[=col_content_compare]{col_content_compare()}}
}
\author{
Sander Devisscher
}
\concept{dataframe_comparison}
