% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_bucket.R
\name{connect_to_bucket}
\alias{connect_to_bucket}
\title{connect to bucket}
\usage{
connect_to_bucket(
  bucket_name,
  bucket_type = "inbo-uat",
  role = "inbo-developers-fis-role"
)
}
\arguments{
\item{bucket_name}{String met de naam van de bucket.}

\item{bucket_type}{String met het type bucket. ofwel "inbo-uat" of "inbo-prod"}

\item{role}{String met de rechtengroep waartoe je behoort. Default is:
"inbo-developers-fis-role"}
}
\value{
Dataframe met een lijst van bestanden op de bucket met relevante info.
}
\description{
Deze functie refresht de sessiontoken nodig om contact te leggen met inbo
AWS S3 buckets. De functie maakt een profiel aan met een sessiontoken die één
uur geldig blijft. Deze heb je nodig om verbinding te maken met de s3 buckets.
}
\section{Voorbereiding:}{
De volgende programma's moeten worden geïnstalleerd vooraleer je deze functie
kunt gebruiken:
-\emph{AWS CLI} moet worden geïnstalleerd als dat nog niet gebeurd is. Hiervoor
heb je admin rechten nodig, een \emph{ict helpdesk call} is dus aan de orde.
Na de installatie moeten je \emph{AWS credentials} eenmalig aangemaakt worden.
Voer hiervoor \verb{aws configure} uit in \emph{windows powershell}. De credentials
kan je bekomen bij Jens Polspoel.
-\emph{devops-toolkit} moet lokaal worden geïnstalleerd dit doe je door
\emph{https://github.com/inbo/devops-toolkit} te clonen dmv \emph{github dekstop}.
Vervolgens kopiëer je \emph{aws-cli-mfa-login.py} & \emph{common.py} naar de Home
Directory van Windows. Default is dat \emph{C:\Users\<voornaam_achternaam>\bin}.
-\emph{Python} moet worden geïnstalleerd als dat nog niet gebeurd is. Dit kan
rechtstreeks in \emph{R} met onderstaande code:
\code{install.packages("reticulate")}
\code{reticulate::install_python()}
\code{system('pip install boto3')}
}

\examples{
 \dontrun{
# Alleen de bucket_name is uiterst noodzakelijk:
source("./src/connect_to_bucket.R")
connect_to_bucket("inbo-exotenportaal-uat-eu-west-1-default")
}
}
\author{
Sander Devisscher

Jens Polspoel
}
