% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-CRS.R
\docType{class}
\name{CRS-class}
\alias{CRS-class}
\title{CRS class definition}
\source{
sp version 2.1-3 by Edzer Pebesma, Roger Bivand
}
\description{
This class is used to store coordinate reference system information for
spatial objects. It is a simple class with a single slot, \code{projargs},
which is a character vector containing the PROJ.4 string representation of
the CRS.
}
\section{Slots}{

\describe{
\item{\code{projargs}}{A character vector containing the PROJ.4 string representation
of the CRS.}
}}

\author{
Copyright (c) 2003-7 by Barry Rowlingson and Roger Bivand
}
\keyword{classes}
