% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRS-methods.R
\name{CRS}
\alias{CRS}
\title{CRS methods}
\source{
sp version 2.1-3 by Edzer Pebesma, Roger Bivand
}
\usage{
CRS(
  projargs = NA_character_,
  doCheckCRSArgs = TRUE,
  SRS_string = NULL,
  get_source_if_boundcrs = TRUE,
  use_cache = TRUE
)
}
\description{
Methods for the CRS class.
}
\author{
Copyright (c) 2003-23 by Barry Rowlingson and Roger Bivand
}
