% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dep_media.R
\name{download_dep_media}
\alias{download_dep_media}
\title{Download deployment media}
\usage{
download_dep_media(
  dataset,
  depID,
  species = NULL,
  favorite = FALSE,
  outputfolder = NULL
)
}
\arguments{
\item{dataset}{character string, path to the folder where a camptraptor datapackage has been unzipped.}

\item{depID}{character string, ID of the deployment to download media from.}

\item{species}{character string, latin name of the species to download}

\item{favorite}{boolean, do you only want the pretty pictures?}

\item{outputfolder}{character string, path where the function should download the media into}
}
\value{
Downloads the specified media files into the outputfolder
}
\description{
This function allows the user to download all media related to a Agouti -
dataset which matches the given parameters.
}
\details{
If you are getting an Authorization Error (#403), this probably means your Agouti project has Restrict Images on. This needs to be turned off.
If depID = "all" and favorite = TRUE, the function will download all favorited pictures in the whole dataset.
}
\examples{
\dontrun{
drg <- fistools::drg_example

# Situation 1: download whole deployment
download_dep_media(dataset = drg,
                    depID = "96413aa6-5f1f-4dfb-8fab-8f06decc179f")

# Situation 2: download only wanted species
download_dep_media(dataset = drg,
                    depID = "96413aa6-5f1f-4dfb-8fab-8f06decc179f",
                    species = "Dama dama")

# Situation 3: download only favorited species media
download_dep_media(dataset = drg,
                    depID = "96413aa6-5f1f-4dfb-8fab-8f06decc179f",
                    species = "Dama dama",
                    favorite = TRUE)

# Situation 4: download only favorited species media
download_dep_media(dataset = drg,
                    depID = "all",
                    favorite = TRUE)
}

}
\seealso{
Other download:
\code{\link[=download_gdrive_if_missing]{download_gdrive_if_missing()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link[=get_last_modified_zip]{get_last_modified_zip()}},
\code{\link[=read_dmogg]{read_dmogg()}},
\code{\link[=unzip_last_modified_zip]{unzip_last_modified_zip()}}

Other agouti:
\code{\link[=agouti_imager]{agouti_imager()}},
\code{\link[=agouti_validate_ai]{agouti_validate_ai()}},
\code{\link[=depopulate_agouti_imager_seq_done]{depopulate_agouti_imager_seq_done()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link{drg_example}},
\code{\link[=populate_agouti_imager_seq_done]{populate_agouti_imager_seq_done()}},
\code{\link[=redo_non_tracked]{redo_non_tracked()}},
\code{\link[=rename_ct_files]{rename_ct_files()}}
}
\author{
Lynn Pallemaerts

Emma Cartuyvels

Sander Devisscher

Soria Delva
}
\concept{agouti}
\concept{download}
