% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gdrive_if_missing.R
\name{download_gdrive_if_missing}
\alias{download_gdrive_if_missing}
\title{Download gdrive if missing}
\usage{
download_gdrive_if_missing(gfileID, destfile, update_always = FALSE, email)
}
\arguments{
\item{gfileID}{character google file token}

\item{destfile}{character destination filename with extention}

\item{update_always}{optional boolean to trigger a download everytime the
function is run. default is FALSE.}

\item{email}{optional character specifying the users email used to access the
googledrive file.}
}
\value{
If the destination file was missing it is downloaded from the
googledrive.
}
\description{
This function downloads the specified file from google drive if the destination
file does not exist. If it does exist the user will be prompted to download
it again.
}
\details{
Its best practice to provide the email in encrypted form. This can be easily
achieved by adding email as an item in a .renviron file or even beter by using
more robust encryption methods.

When a \emph{PERMISSION_DENIED} error occurs, it is likely that the file is not shared
with the email address provided. This can be fixed by sharing the file with the
email address.
If the file was shared correctly this might indicate that email address
does not have the correct permissions to access the file. This can be fixed by
running \code{googledrive::drive_deauth()} followed by \code{googledrive::drive_auth()} and
making sure the 'show, modify and delete all drive files' option is selected.
Additionally running \code{options(gargle_oauth_cache = ".secrets")} prior to running
the function can fix this issue.
If the error yet persists
please create a \href{https://github.com/inbo/fistools/issues/new/choose}{new issue}
on the github page.
}
\examples{
\dontrun{
# download newest version of the team charter
download_gdrive_if_missing(gfileID = "1gtqcZojPnbLhEgpul3r9sy2zK3UyyCVG",
                           destfile = "../../Teamcharters/Teamcharter_FIS.pdf",
                           email = Sys.getenv("email"),
                           update_always = TRUE)
}
\dontrun{
# download newest DRG Agouti export
download_gdrive_if_missing(gfileID = "1FX8DDyREKMH1M3iW9ijWjVjO_tBH8PXi",
                           destfile = "../fis-projecten/Grofwild/Drongengoed/Input/Agouti/drongengoed_240502.zip",
                           email = Sys.getenv("email"),
                           update_always = TRUE)
}
}
\seealso{
Other download:
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link[=get_last_modified_zip]{get_last_modified_zip()}},
\code{\link[=read_dmogg]{read_dmogg()}},
\code{\link[=unzip_last_modified_zip]{unzip_last_modified_zip()}}
}
\author{
Sander Devisscher
}
\concept{download}
