% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_media.R
\name{download_media}
\alias{download_media}
\title{Download media}
\usage{
download_media(
  dataset,
  location,
  start = NULL,
  end = NULL,
  species = NULL,
  outputfolder
)
}
\arguments{
\item{dataset}{character string, path to the folder where a camptraptor datapackage has been unzipped.}

\item{location}{character string, ID of the cameralocation to download media from}

\item{start}{optional date, startdate of the download. Expected format DD/MM/YYYY.}

\item{end}{optional data, enddate of the download. Expected format DD/MM/YYYY.}

\item{species}{character string, latin name of the species to download}

\item{outputfolder}{character string, path where the function should download the media into}
}
\value{
Downloads the specified media files into the outputfolder
}
\description{
This function allows the user to download all media related to a Agouti -
dataset which matches the given parameters.
}
\details{
If you are getting an Authorization Error (#403), this probably means your Agouti project has Restrict Images on. This needs to be turned off.
}
\examples{
\dontrun{
# unzip camtraptor data package
unzip("./Grofwild/Drongengoed/Input/Agouti/drongengoed_230601.zip",
     exdir = "./Grofwild/Drongengoed/Input/Agouti/drongengoed")
drg <- camtraptor::read_camtrap_dp(file = "./Grofwild/Drongengoed/Input/Agouti/drongengoed")

# Situation 1: download whole deployment
download_media(dataset = drg,
               location = "D98_427",
               start = "16/12/2022",
               outputfolder = "C:/Users/lynn_pallemaerts/Pictures") # If somebody else is testing this: change to your user name

# Situation 2: download only wanted species
download_media(dataset = drg,
               location = "D121_423",
               start = "25/03/2023",
               species = "Capreolus capreolus",
               outputfolder = "C:/Users/lynn_pallemaerts/Pictures") # If somebody else is testing this: change to your user name

download_media(dataset = drg,
               location = "D55_340",
               end = "27/05/2023",
               species = "Lepus europaeus",
               outputfolder = "C:/Users/lynn_pallemaerts/Pictures") # If somebody else is testing this: change to your user name

# Situation 3: Try to download species that is not present
download_media(dataset = drg,
               location = "D55_340",
               end = "27/05/2023",
               species = "Dama dama",
               outputfolder = "C:/Users/lynn_pallemaerts/Pictures") # If somebody else is testing this: change to your user name

# cleanup after use
remove(drg)
unlink("./Grofwild/Drongengoed/Input/Agouti/drongengoed", recursive = TRUE)
}

}
\author{
Lynn Pallemaerts

Emma Cartuyvels

Sander Devisscher
}
