% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wekeo_data.R
\name{earthkit_download}
\alias{earthkit_download}
\title{Wrapper function for ekd.from_source}
\usage{
earthkit_download(
  source = "wekeo",
  dataset_id,
  api_request,
  cache_dir = file.path(Sys.getenv("USERPROFILE"), ".earthkit_cache")
)
}
\arguments{
\item{source}{The data source, e.g., "wekeo".}

\item{dataset_id}{The dataset identifier.}

\item{api_request}{The API request in JSON format.}

\item{cache_dir}{Directory to store cached files.}
}
\value{
The path to the downloaded file.
}
\description{
Wrapper function for ekd.from_source
}
