% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localhost_connection.R
\name{end_localhost_connection}
\alias{end_localhost_connection}
\title{Close localhost connection}
\usage{
end_localhost_connection(port = 5555)
}
\arguments{
\item{port}{connection port of the localhost instance, default = 5555}
}
\description{
A local host connection will remain open unless closed therefor you should
use this function to close it. This function uses \code{get_active_localhost}
under the hood to collect a list of active localhost connections on a given
port.
}
\seealso{
\code{\link[=start_localhost_connection]{start_localhost_connection()}}
}
\author{
Jens Polspoel

Sander Devisscher
}
