% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agouti_validate_ai.R
\name{find_ai}
\alias{find_ai}
\title{Find used AI models in a datapackage}
\usage{
find_ai(datapack, species = species)
}
\arguments{
\item{datapack}{a camtrapdb datapackage containing the column classificationMethod}

\item{species}{a vector of species passed down from \code{agouti_validate_ai}.}
}
\value{
a vector of ai models used
}
\description{
Find used AI models in a datapackage
}
