% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_to_drive.R
\name{folder_to_drive}
\alias{folder_to_drive}
\title{Upload a folder to Google Drive}
\usage{
folder_to_drive(
  input_folder = NULL,
  drive_folder = NULL,
  zipfiles = FALSE,
  cleanup = FALSE,
  filetype = "all",
  recursive = FALSE,
  email = Sys.getenv("email")
)
}
\arguments{
\item{input_folder}{character string, path to the folder you want to upload}

\item{drive_folder}{character string, id to the folder in Google Drive where you want to upload the files.}

\item{zipfiles}{boolean, whether to zip the files before uploading. Default is FALSE.}

\item{cleanup}{boolean, whether to delete the local files after uploading. Default is FALSE.}

\item{filetype}{character string, the type of files to upload. Default is "all". You can specify a filetype by using a regular expression, for example "csv" to upload only csv files.}

\item{recursive}{boolean, whether to include files in subfolders of the input folder. Default is FALSE.}

\item{email}{character string, the email address to use for authentication with Google Drive. Default is the value of the "email" environment variable.}
}
\value{
The specified files are uploaded to Google Drive, and optionally the local files are deleted.
}
\description{
This function allows you to upload a folder to Google Drive. You can choose to zip the folder before uploading, and to clean up the local files after uploading.
}
\details{
You can find the id of a folder in Google Drive by right-clicking on the folder, selecting "Get link", and copying the part of the URL that comes after "folders/". For example, if the URL is "https://drive.google.com/drive/folders/1a2b3c4d5e6f7g8h9i0j", then the id of the folder is "1a2b3c4d5e6f7g8h9i0j".
If you choose to zip the folder, only the zip file will be uploaded to Google Drive. If you choose to clean up the local files, all files in the input folder will be deleted after uploading.
}
\examples{
\dontrun{
# upload a folder to Google Drive without zipping and without cleaning up local files
folder_to_drive(input_folder = "path/to/your/folder",
                drive_folder = "1a2b3c4d5e6f7g8h9i0j",
                zipfiles = FALSE,
                cleanup = FALSE,
                filetype = "all",
                recursive = FALSE,
                email = Sys.getenv("email"))

# upload a folder to Google Drive with zipping and cleaning up local files, only for csv files
folder_to_drive(input_folder = "path/to/your/folder",
               drive_folder = "1a2b3c4d5e6f7g8h9i0j",
               zipfiles = TRUE,
               cleanup = TRUE,
               filetype = "csv",
               recursive = TRUE,
               email = Sys.getenv("email"))
}

}
\seealso{
Other download: 
\code{\link{download_dep_media}()},
\code{\link{download_gdrive_if_missing}()},
\code{\link{download_seq_media}()},
\code{\link{get_last_modified_zip}()},
\code{\link{unzip_last_modified_zip}()}
}
\author{
Sander Devisscher
}
\concept{download}
