% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_modified_zip.R
\name{get_last_modified_zip}
\alias{get_last_modified_zip}
\title{get last modified zip}
\usage{
get_last_modified_zip(folder_id)
}
\arguments{
\item{folder_id}{The ID of the googledrive folder to check for the last
modified zip file.}
}
\value{
The name of the last modified zip file in the specified folder.
}
\description{
A helper function to retrieve the id of last modified zip file from a specified
Google Drive folder.
}
\examples{
 \dontrun{
 get_last_modified_zip(folder_id = "17p2MZt9LIuhIU72u_JjDBbO7D1IPTv7-")
}

}
\seealso{
Other download:
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_gdrive_if_missing]{download_gdrive_if_missing()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link[=read_dmogg]{read_dmogg()}},
\code{\link[=unzip_last_modified_zip]{unzip_last_modified_zip()}}
}
\author{
Sander Devisscher, Martijn Bollen
}
\concept{download}
