% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wekeo_data.R
\name{get_wekeo_data}
\alias{get_wekeo_data}
\title{Get Wekeo data using Earthkit}
\usage{
get_wekeo_data(
  dataset_id,
  api_request,
  bbox = NULL,
  productType,
  resolution,
  startdate,
  enddate,
  itemsPerPage = 200,
  startIndex = 0
)
}
\arguments{
\item{dataset_id}{The dataset identifier.}

\item{api_request}{The API request in JSON format.}

\item{bbox}{A bounding box for the data request (optional).}

\item{productType}{The type of product to request (optional).}

\item{resolution}{The resolution of the data (optional).}

\item{startdate}{The start date for the data request (optional).}

\item{enddate}{The end date for the data request (optional).}

\item{itemsPerPage}{The number of items per page (default is 200).}

\item{startIndex}{The starting index for pagination (default is 0).}
}
\value{
A terra SpatRaster object containing the requested data.
}
\description{
Get Wekeo data using Earthkit
}
\details{
This function retrieves data from the \href{https://wekeo.copernicus.eu/}{Wekeo service} using the Earthkit library.
It constructs an API request based on the provided parameters and downloads the data.
This function is a wrapper around the \code{earthkit_download} from python function.
To get this function to work, you need to have the \code{earthkit} python package installed
which depends on the \verb{Microsoft C++ Build Tools}.
This toolkit is required to compile the \code{earthkit} package.
To install the toolkit please place ict helpdesk call at \code{ict.helpdesk@inbo.be}

When no API request is provided the function constructs a custom API request.
When the API request is provided, it will be used directly, ignoring the other parameters (except dataset_id).
}
\examples{
\dontrun{

# Example usage with a custom API request
api_request <- '{
"dataset_id": "EO:CLMS:DAT:CLMS_GLOBAL_NDVI_300M_V1_10DAILY_NETCDF",
"productType": "NDVI300",
"resolution": "300",
"startdate": "2021-01-01T00:00:00.000Z",
"enddate": "2021-01-29T23:59:59.999Z",
"itemsPerPage": 200,
"startIndex": 0
}'

Call the function with the custom API request
data <- get_wekeo_data(
  dataset_id = "EO:CLMS:DAT:CLMS_GLOBAL_NDVI_300M_V1_10DAILY_NETCDF",
  api_request = api_request
  )
}
}
\author{
Martijn Bollen & Sander Devisscher

This function retrieves data from the Wekeo service using the Earthkit library.
}
