% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_intersect.R
\name{get_wfs_layers}
\alias{get_wfs_layers}
\title{Function to get the available layers in a wfs}
\usage{
get_wfs_layers(url)
}
\arguments{
\item{url}{A character string with the url of the wfs}
}
\value{
A character vector with the available layers in the wfs
}
\description{
Function to get the available layers in a wfs
}
