% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_depricated.R
\name{install_depricated}
\alias{install_depricated}
\title{install depricated package}
\usage{
install_depricated(package_name, version, force = FALSE)
}
\arguments{
\item{package_name}{A character string indicating the name of the package to be installed}

\item{version}{A character string indicating the version of the package to be installed}

\item{force}{A logical indicating whether the installation should be forced}
}
\description{
Helper function that installs depricated package when missing, from a tarball.
}
\details{
Sometimes a package is depricated and no longer available on CRAN.
This function allows you to install a depricated package from a tarball.
It is recommended to use this function only when necessary.
In first instance try to find an alternative package.
Using depricated packages may cause conflicts and unexpected behaviour.
}
\examples{
\dontrun{
# install latest version of sp package instead of install_sp()
install_depricated("sp", "2.1-3")
# trying to install a wrong version of maptools
install_depricated("maptools", "2.1-3", force = TRUE)

}

}
\seealso{
Other other: 
\code{\link{install_sp}()},
\code{\link{label_selecter}()}
}
\author{
Sander Devisscher
}
\concept{other}
