% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sp.R
\name{install_sp}
\alias{install_sp}
\title{install sp}
\usage{
install_sp(force = FALSE)
}
\arguments{
\item{force}{A logical indicating whether the installation should be forced}
}
\description{
Helper function that installs sp when missing, from a tarball.
}
\details{
The "sp" package should be unloaded from the namespace when it is not needed anymore.
Every function that uses "sp" should start with a call to this function.
And end with a call to \code{unloadNamespace("sp")}.
This is because the "sp" package is known to cause conflicts with other packages.
}
\author{
Sander Devisscher
}
