% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_crs.R
\docType{data}
\name{lib_crs}
\alias{lib_crs}
\title{Lib CRS}
\format{
A named list with the following elements:
\describe{
\item{CRS_Naam}{ID of the CRS}
\item{Proj4s}{Javascript Object Notation (JSON) of the CRS}
\item{ESPG}{EPSG code of the CRS}
}
}
\source{
WGS \url{https://epsg.io/4326}

BEL72 \url{https://epsg.io/31300}
}
\usage{
lib_crs
}
\description{
Library of commonly used coordinate reference system (CRS) codes
}
\concept{library}
\keyword{datasets}
