% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_line_endings.R
\name{normalize_line_endings}
\alias{normalize_line_endings}
\title{Normalize line endings according to \code{.gitattributes}}
\usage{
normalize_line_endings()
}
\value{
Files in the directory will be altered where relevant. Users get the option to commit these changes.
}
\description{
This function adjusts the line endings (LF or CRLF) and enforces file attribute handling of all files in the current directory and subdirectories, according to the settings specified in the \code{.gitattributes} file.
This creates consistency between the files on the online repository and those in the local environment.
}
\details{
The function
\itemize{
\item Checks whether a \code{.gitattributes} file is available.
\item Normalizes all files using Git.
\item Checks if files were altered and asks users if these changes can be committed.
\item Provides status updates, so users can follow the progress of the function.
}
}
\examples{
\dontrun{
# Run normalization and (optionally) commit changes
normalize_line_endings()
}
}
\author{
Soria Delva
}
