% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agouti_imager_helpers.R
\name{populate_agouti_imager_seq_done}
\alias{populate_agouti_imager_seq_done}
\title{Populate agouti_imager_seq_done}
\usage{
populate_agouti_imager_seq_done(
  dir = NULL,
  seqID = NULL,
  email = Sys.getenv("email"),
  sheet_id = "1PcqJziXm-ZNbCi2JJliQH_FQY8YMPXNEGgYwiiP2Ws8"
)
}
\arguments{
\item{dir}{a character string specifying the directory to look for \code{seq_done.csv}}

\item{seqID}{a character list of sequences to add}

\item{email}{a character specifying your email used to authenticate googlesheet}

\item{sheet_id}{the character string of the \href{https://docs.google.com/spreadsheets/d/1PcqJziXm-ZNbCi2JJliQH_FQY8YMPXNEGgYwiiP2Ws8/edit?gid=0#gid=0}{agouti_imager_seq_done} googlesheet}
}
\description{
A script to retrieve locally available seq_done files upload the sequences
to the \href{https://docs.google.com/spreadsheets/d/1PcqJziXm-ZNbCi2JJliQH_FQY8YMPXNEGgYwiiP2Ws8/edit?gid=0#gid=0}{agouti_imager_seq_done} googlesheet.
}
\details{
If both dir and seqID are provided the function will only add the seqIDs.
}
\examples{
\dontrun{

seqID <- c("f9cfa2ac-36a5-45e0-bd84-2b3a5a234ee6",
"793b5057-90b3-47db-b400-96b2b31c9b59",
"53075b53-dd7b-41b0-b778-394f28e16268",
 "560c8d87-ec9a-43f2-adb9-d96a09f589ef",
  "7b2d6285-34af-4bd3-b9f9-40022a267ad8")

populate_agouti_imager_seq_done(seqID = seqID)
}
}
\seealso{
Other agouti:
\code{\link[=agouti_imager]{agouti_imager()}},
\code{\link[=agouti_validate_ai]{agouti_validate_ai()}},
\code{\link[=depopulate_agouti_imager_seq_done]{depopulate_agouti_imager_seq_done()}},
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link{drg_example}},
\code{\link[=redo_non_tracked]{redo_non_tracked()}},
\code{\link[=rename_ct_files]{rename_ct_files()}}
}
\concept{agouti}
