% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCI.R
\name{qd_pci2_D}
\alias{qd_pci2_D}
\title{Distance matrix for qd_pci2}
\usage{
qd_pci2_D(x, m = 1, p = 1)
}
\arguments{
\item{x}{vector with the scores of the respondents}

\item{m}{m value in the formula (see details)}

\item{p}{power value in the formula (see details)}
}
\value{
single value containing pci index
}
\description{
Calculates distance matrix for the function qd_pci2
}
\details{
\deqn{Dp_{x,y} = (|r_{x} - r_{y}|) - (m-1))^{p}}
\deqn{if sign(r_{x} \neq r_{y}) \\
else d_{x,y} = 0}
Dp_x,y = (|r_x - r_y| - (m-1))^p
}
\examples{
\dontrun{
#'set.seed(201)
Xv <- sample(-2:2, size = 100, replace = TRUE) #random responses
qd_pci2(Xv, scale_values = -2:2, x_is_table = FALSE, m = 1, p = 1) # 0.37
}
}
\seealso{
Other plotting:
\code{\link[=qd_pci1]{qd_pci1()}},
\code{\link[=qd_pci2]{qd_pci2()}}
}
\concept{plotting}
