% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dmogg.R
\name{read_dmogg}
\alias{read_dmogg}
\title{Read "dieren met onderkaakgegevens georef"}
\usage{
read_dmogg(type = "short", email = Sys.getenv("email"))
}
\arguments{
\item{type}{"short" or "long" indicating which version you want to import.
The default is "short". See details for more info.}

\item{email}{email used to authenticate the google drive}
}
\value{
a dataframe containing the most recent DMOGG data.
}
\description{
This function reads the most up to date version of the
dieren met onderkaakgegevens georef (DMOGG) file from the backoffice wild
analyse.
}
\details{
Two versions of the DMOGG dataset can be imported by this function, namely:
\itemize{
\item The "Short" version containing only data collected via the e-loket and later
WIZ-app from ANB enriched with data from the autopsiedb.
The temporal range of this dataset is 2014 - today.
This dataset is more complete and contains spatial data concerning wbe's & fbz's.
\item The "Long" version containing the data from the "Short" dataset plus data
collected via the predecessors of the e-loket (ao zwijntje) by INBO.
This dataset is less complete because early data (<2014) misses almost all spatial info.
}

This function uses \code{fistools::download_gdrive_if_missing()} under the hood to
download the file from the google drive.
}
\examples{
\dontrun{
# read the most recent short DMOGG data
dmogg <- read_dmogg()
}

}
\seealso{
Other download:
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_gdrive_if_missing]{download_gdrive_if_missing()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link[=get_last_modified_zip]{get_last_modified_zip()}},
\code{\link[=unzip_last_modified_zip]{unzip_last_modified_zip()}}
}
\author{
Sander Devisscher
}
\concept{download}
