% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_ct_files.R
\name{rename_ct_files}
\alias{rename_ct_files}
\title{Rename camera trap files for upload in Agouti}
\usage{
rename_ct_files(
  foldername,
  extensions = c("jpg", "png", "jpeg", "bmp"),
  compile = FALSE
)
}
\arguments{
\item{foldername}{DCIM folder to be treated}

\item{extensions}{file extensions to be rename. Default is jpg, jpeg, bmp and png}

\item{compile}{logical. If TRUE, all files will be compiled in a single folder}
}
\value{
renamed picture files in the original folder
}
\description{
Function that rename camera trap files (i.e. photos) by
appending folder name (e.g. 1000RECX) to file name (e.g. IMG0001) to create
all unique file names across a deployment. Needed for deployments with more
than 9999 pictures, so they can be uploaded in Agouti. This is a copy of the
old \code{rename_file_names()} in the old \code{fis-projecten} (archived).
}
\details{
This function will rename all files in the subfolders of the folder specified
by foldername. If no foldername is provided, the user will be prompted to
select a folder. The function will rename all files with the extensions
specified in the extensions argument. The function will also compile all files
in a single folder if compile is set to TRUE.
}
\examples{
\dontrun{
tempzip <- tempfile(fileext = ".zip")
fistools::download_gdrive_if_missing(gfileID = "1-84hbKouLKGwnLgBSRaQO1BnfogoFZWz",
                                     destfile = tempzip,
                                     email = Sys.getenv("email"),
                                     update_always = TRUE)

foldername <- paste0(tempdir(), "/test_case_renaming")
unzip(tempzip,
      exdir = tempdir())
rename_ct_files(foldername)
browseURL(foldername)
unlink(foldername,
       recursive = TRUE)

# The function also works when no foldername is provided
rename_ct_files()

# The function can also compile all files in a single folder
rename_ct_files(foldername, compile = TRUE)
browseURL(foldername)
unlink(foldername,
       recursive = TRUE)
}

}
\seealso{
Other agouti:
\code{\link[=agouti_imager]{agouti_imager()}},
\code{\link[=agouti_validate_ai]{agouti_validate_ai()}},
\code{\link[=depopulate_agouti_imager_seq_done]{depopulate_agouti_imager_seq_done()}},
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link{drg_example}},
\code{\link[=populate_agouti_imager_seq_done]{populate_agouti_imager_seq_done()}},
\code{\link[=redo_non_tracked]{redo_non_tracked()}}
}
\author{
: Lynn Pallemaerts and Someone Before Her
}
\concept{agouti}
