% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_ct_files.R
\name{rename_file_names}
\alias{rename_file_names}
\title{Rename camera trap files for upload in Agouti}
\usage{
rename_file_names(foldername, extensions = c("jpg", "png", "jpeg", "bmp"))
}
\arguments{
\item{foldername}{DCIM folder to be treated}

\item{extensions}{file extensions to be rename. Default is jpg, jpeg, bmp and png}
}
\value{
renamed picture files in the original folder
}
\description{
Function that rename camera trap files (i.e. photos) by
appending folder name (e.g. 1000RECX) to file name (e.g. IMG0001) to create
all unique file names across a deployment. Needed for deployments with more
than 9999 pictures, so they can be uploaded in Agouti. This is a copy of the
old \code{rename_file_names()} in the old \code{fis-projecten} (archived).
}
\examples{
\dontrun{
foldername <- "./data/test_case_renaming/"
rename_file_names(foldername)
}
#'@export
}
\author{
: Lynn Pallemaerts and Someone Before Her
}
