% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_function.R
\name{retry_function}
\alias{retry_function}
\title{Retry a function multiple times}
\usage{
retry_function(expr, max_attempts = 3, sleep_time = 0)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{max_attempts}{The maximum number of attempts to make.}

\item{sleep_time}{The time to sleep between attempts, in seconds.}
}
\value{
The result of the expression if successful.
}
\description{
This function evaluates an expression multiple times until it succeeds or the
maximum number of attempts is reached.
}
\examples{
\dontrun{
# This example will fail or succeed randomly.

some_function <- function() {
 if (runif(1) < 0.5) {  # Randomly fail
   stop("\U02620 Something went wrong!")
   }
 return("Success")
}

retry_function({some_function()}, max_attempts = 5, sleep_time = 1)
}

}
\seealso{
Other other:
\code{\link[=label_selecter]{label_selecter()}}
}
\author{
Sander Devisscher
}
\concept{other}
