% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wekeo_data.R
\name{setup_hda_credentials}
\alias{setup_hda_credentials}
\title{Setup HDA credentials for accessing Wekeo data}
\usage{
setup_hda_credentials()
}
\value{
A Python client object for accessing Wekeo data.
}
\description{
This function sets up the HDA (Harmonized Data Access) credentials required to access Wekeo data.
It checks if the HDA Python module is available, installs it if necessary,
and prompts the user for their HDA username and password if they are not already stored in the \code{.hdarc} file.
}
\examples{
\dontrun{
# Setup HDA credentials
hda_client <- setup_hda_credentials()
}

}
