% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_to_geojson.R
\name{shape_to_geojson}
\alias{shape_to_geojson}
\title{Convert shapefiles to geojson
Deze functie zet alle shapes in een specifieke map, één bestand of een lijst van bestanden
van .shp om naar .geojson.
Daarnaast wordt de projectie getransformeerd naar wgs84 of een andere projectie.}
\usage{
shape_to_geojson(input, output, output_crs = 4326, overwrite = "ask")
}
\arguments{
\item{input}{een character string, een map, een bestand of een lijst van bestanden}

\item{output}{een character string, de map waar de geojson bestanden worden
opgeslagen, default is de input map}

\item{output_crs}{een integer, de projectie van de output, default is wgs84}

\item{overwrite}{een boolean of character string, Vraag of de bestanden mogen
overschreven worden, default is "ask"}
}
\description{
Convert shapefiles to geojson
Deze functie zet alle shapes in een specifieke map, één bestand of een lijst van bestanden
van .shp om naar .geojson.
Daarnaast wordt de projectie getransformeerd naar wgs84 of een andere projectie.
}
\details{
De functie checkt of de input een map is, als dit het geval is worden alle .shp
bestanden in de map omgezet naar .geojson. Als de input geen map is, wordt de
input als bestand beschouwd en wordt deze omgezet naar .geojson.

Als de output niet is gespecificeerd, wordt de output gelijkgesteld aan de input.
}
\examples{
\dontrun{
# Voorbeeld van hoe de shape_to_geojson functie te gebruiken
# Sla boswachterijen_2024 op als .shp bestand in een tempdir
boswachterijen_2024 <- fistools::boswachterijen$boswachterijen_2024
tempdir <- tempdir()
sf::st_write(boswachterijen_2024, paste0(tempdir, "/boswachterijen_2024.shp"))

# controleer of de shp goed opgeslagen werd
browseURL(tempdir)

# Zet de shp om naar geojson
shape_to_geojson(input = tempdir)

# Read and plot the geojson
boswachterijen_2024_geojson <- sf::st_read(paste0(tempdir, "/boswachterijen_2024.geojson"))
leaflet::leaflet() \%>\%
 leaflet::addTiles() \%>\%
 leaflet::addPolygons(data = boswachterijen_2024_geojson)
 }
}
\seealso{
Other spatial: 
\code{\link{CRS_extracter}()},
\code{\link{apply_grtsdb}()},
\code{\link{calculate_polygon_centroid}()},
\code{\link{collect_osm_features}()}
}
\concept{spatial}
