% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localhost_connection.R
\name{start_localhost_connection}
\alias{start_localhost_connection}
\title{Start localhost connection}
\usage{
start_localhost_connection(
  db = "2",
  envir = "inbo-prod",
  role = "inbo-autopsies-rds-connect-role",
  devops_toolkit_path = "../../../bin/devops-toolkit-windows-amd64.exe",
  port = 5555
)
}
\arguments{
\item{db}{the number corresponding to the db you want to access, default is 2}

\item{envir}{the environment in which the db resides, default is "inbo-prod"}

\item{role}{your role used to connect to the db}

\item{devops_toolkit_path}{the path, ending with \code{.exe}, where the devops-toolkit
was installed.}

\item{port}{connection port of the localhost instance, default = 5555}
}
\description{
After the upgrade to a newer version of postgresql connecting to the
autopsiedb no longer occurs via the bastion but via a modified localhost.
This function allows connection to newer postgresql databases. But its mainly
focused on the autopsiedb.
}
\section{Voorbereiding:}{
De volgende programma's moeten worden geïnstalleerd vooraleer je deze functie
kunt gebruiken:
-\emph{AWS CLI} moet worden geïnstalleerd als dat nog niet gebeurd is. Hiervoor
heb je admin rechten nodig, een \emph{ict helpdesk call} is dus aan de orde.
Na de installatie moeten je \emph{AWS credentials} eenmalig aangemaakt worden.
Voer hiervoor \verb{aws configure} uit in \emph{windows powershell}. De credentials
kan je bekomen bij Jens Polspoel.
-\emph{devops-toolkit} moet lokaal worden geïnstalleerd dit doe je door de laatste
versie van \emph{https://github.com/inbo/devops-toolkit/releases/tag/ (>= v1.0.3)}
te downloaden. De functie werkt enkel met de windows versie van devops-toolkit

Vervolgens kopiëer je \emph{aws-cli-mfa-login} naar de Home
Directory van Windows. Default is dat \emph{C:/Users/\%voornaam_achternaam\%/bin}.
Hernoem het bestand vervolgens naar \emph{aws-cli-mfa-login.exe}.
}

\examples{
\dontrun{
 start_localhost_connection()

 con <- odbc::dbConnect(odbc::odbc(),
                      Driver = "PostgreSQL Unicode(x64)",
                      database = "autopsies",
                      server = "127.0.0.1",
                      port = 5555,
                      uid = Sys.getenv("autopsies_user"),
                      pwd = paste0("{", Sys.getenv("autopsies_pwd"), "}"),
                      sslmode = "require")

 identificatie <- DBI::dbGetQuery(con,
                                  "SELECT * FROM public.identificatie")

end_localhost_connection()
}

}
\seealso{
\code{\link[=end_localhost_connection]{end_localhost_connection()}}
}
\author{
Jens Polspoel

Sander Devisscher
}
