% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunsetter.R
\name{sunsetter}
\alias{sunsetter}
\title{sunsetter
Calculate the sunrise and sunset times for a given range of dates and location}
\usage{
sunsetter(
  StartDate = Sys.Date(),
  EndDate = Sys.Date(),
  lat = 50.866572,
  lng = 4.350309
)
}
\arguments{
\item{StartDate}{Date in \%y-\%m-\%d format indicating startdate of dataframe. Defaults to today}

\item{EndDate}{Date in \%y-\%m-\%d format indicating enddate of dataframe. Defaults to today}

\item{lat}{Numeric indicating the latitude. Defaults to Herman Teirlinck building in Brussels}

\item{lng}{Numeric indicating the longitude. Defaults to Herman Teirlinck building in Brussels}
}
\value{
dataframe containing the dates between the startdate and enddate, the corresponding sunrise time and sunset time.
}
\description{
sunsetter
Calculate the sunrise and sunset times for a given range of dates and location
}
\details{
This function uses the sunrise-sunset API to calculate the sunrise and sunset
times for a given range of dates and fixed location.
The default location is the Herman Teirlinck building in Brussels.

if StartDate and EndDate are not specified, the function will return the sunrise
and sunset times for today.
}
\examples{
\dontrun{
# sunrise and sunset times for the first 10 days of 2021 in Brussels
sunsetter(StartDate = "2021-01-01", EndDate = "2021-01-10", lat = 50.866572, lng = 4.350309)

# sunrise and sunset times for today in Brussels
sunsetter()
}

}
\seealso{
Other temporal:
\code{\link[=sunsetter2]{sunsetter2()}}
}
\author{
Sander Devisscher
}
\concept{temporal}
