% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyMonkey_q6.R
\docType{data}
\name{surveyMonkey_q6}
\alias{surveyMonkey_q6}
\title{surveyMonkey_q6}
\format{
\subsection{\code{surveyMonkey_q6}}{

Een data frame met 1287 rijen en 5 kolommen:
\describe{
\item{respondent_id}{Unieke ID per respondent}
\item{question}{Beschrijving van de vraag}
\item{topic}{Beschrijving van het onderwerp van de vraag}
\item{response}{Beschrijving van het antwoord op de vraag}
\item{score}{De score op Likert-schaal die overenkomt met het antwoord op de vraag}
...
}
}
}
\source{
\url{https://github.com/inbo/big-picture-m5/blob/main/Interim/question6.csv}
}
\usage{
surveyMonkey_q6
}
\description{
Vraag en antwoorden van respondenten op vraag 6 van de survey Monkey bevraging,
uitgevoerd in het kader van WP3 van het Big picture project.
}
\seealso{
Other dataset:
\code{\link{CLC18_BE}},
\code{\link{boswachterijen}},
\code{\link{drg_example}}
}
\concept{dataset}
\keyword{datasets}
