% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_modified_zip.R
\name{unzip_last_modified_zip}
\alias{unzip_last_modified_zip}
\title{Unzip last modified zip}
\usage{
unzip_last_modified_zip(folder_id, exdir = tempdir())
}
\arguments{
\item{folder_id}{The ID of the googledrive folder to check for the last modified zip file.}

\item{exdir}{The directory where the zip file should be extracted. Defaults to \code{tempdir()}.}
}
\value{
The path to the datapackage.json of the last modified zip file in the specified folder.
}
\description{
A function to download and unzip the last modified zip file from a specified
Google Drive folder.
}
\examples{
\dontrun{
datapackage <- unzip_last_modified_zip(folder_id = "17p2MZt9LIuhIU72u_JjDBbO7D1IPTv7-") |>
camtraptor::read_camtrap_dp()
}

}
\seealso{
Other download:
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_gdrive_if_missing]{download_gdrive_if_missing()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link[=get_last_modified_zip]{get_last_modified_zip()}},
\code{\link[=read_dmogg]{read_dmogg()}}
}
\concept{download}
