% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUID_List.R
\name{UUID_List}
\alias{UUID_List}
\title{UUID list generator}
\usage{
UUID_List(temp_input)
}
\arguments{
\item{temp_input}{a data.frame to which UUIDs should be appended}
}
\description{
A helper script to generate a list of UUIDs
}
