% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_intersect.R
\name{wfs_intersect}
\alias{wfs_intersect}
\title{Extract values from web feature service (wfs)}
\usage{
wfs_intersect(df, x_lam, y_lam, url, layer)
}
\arguments{
\item{df}{A \code{data.frame} with x and y coordinates in the Belgian Lambert 72 (EPSG:31370)}

\item{x_lam}{A character string with the column name of the X coordinate}

\item{y_lam}{A character string with the column name of the Y coordinate}

\item{url}{A character string with the url of the wfs}

\item{layer}{A character string with the name of the layer in the wfs}
}
\value{
A \code{data.frame} with the values of the wfs appended to the list of points
}
\description{
This function extracts values from a web feature service (wfs)
}
\examples{
\dontrun{
}
}
\concept{wfs_functions}
