% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agouti_validate_ai.R
\name{agouti_validate_ai}
\alias{agouti_validate_ai}
\title{Validate AI classifications in Agouti}
\usage{
agouti_validate_ai(
  gfileID,
  ai_model = "Europe",
  agouti_prj_id,
  species = NULL,
  email
)
}
\arguments{
\item{gfileID}{The Google Drive file ID of the datapackage to be downloaded.}

\item{ai_model}{The name of the AI model used for classification (e.g., "Europe"). Default is "Europe".}

\item{agouti_prj_id}{The Agouti project ID where the sequences will be opened for validation.}

\item{species}{Optional. A vector of species (scientific names) to filter the
observations for validation. If NULL, the user will be prompted to select species
from the available options.}

\item{email}{Optional. The email address used for authentication when downloading the
datapackage from Google Drive. If not provided, the function will attempt to
retrieve it from system environment variables or prompt the user for input.}
}
\value{
None. The function opens URLs in the default web browser for validation.
}
\description{
This function allows you to validate AI classifications in Agouti by
downloading a datapackage from Google Drive, filtering observations based on
the specified AI model and species, and opening the relevant sequence IDs in
Agouti for manual validation.
}
\details{
The function performs the following steps:
\enumerate{
\item Checks for the required parameters and prompts the user if necessary.
\item Downloads the specified datapackage from Google Drive using the provided file ID.
\item Unzips the downloaded datapackage and reads it using the \code{camtraptor} package.
\item Filters the observations based on the specified AI model and species.
\item Extracts the unique sequence IDs from the filtered observations and opens them in Agouti for manual validation using the \code{agouti_imager} function.
\item The user can validate the classifications in Agouti
\emph{NOTE:} This function doesn't track which sequences have been validated!
}

Filtering occurs using \code{base::grepl()} on the \code{classifiedBy} field.
This means that when \code{ai_model = "Europe"}, the sequences with
\code{classifiedBy \%in\% c("Europe v1", "Europe v1.2", "Europe v2", ...)} are
withheld to be manually validated using \code{agouti_imager}.
}
\examples{
\dontrun{
validate_agouti_ai(gfileID = "your_google_drive_file_id",
                   ai_model = "Europe",
                   agouti_prj_id = "your_agouti_project_id",
                   species = c("Lynx lynx", "Canis lupus"))
}


}
\seealso{
Other agouti:
\code{\link[=agouti_imager]{agouti_imager()}},
\code{\link[=depopulate_agouti_imager_seq_done]{depopulate_agouti_imager_seq_done()}},
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link{drg_example}},
\code{\link[=populate_agouti_imager_seq_done]{populate_agouti_imager_seq_done()}},
\code{\link[=redo_non_tracked]{redo_non_tracked()}},
\code{\link[=rename_ct_files]{rename_ct_files()}}
}
\author{
Sander Devisscher
}
\concept{agouti}
