% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boswachterijen.R
\docType{data}
\name{boswachterijen}
\alias{boswachterijen}
\title{Boswachterijen}
\format{
\subsection{\code{boswachterijen}}{

En sf data frame with 98 rijen and 11 kolommen per jaar:
\describe{
\item{Regio}{Beheerregio}
\item{Naam}{Naam van de boswachter}
\item{telefoon}{Telefoon nr van de boswachter}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
boswachterijen
}
\description{
Spatiale en andere informatie (o.a. telefoonummers) van de boswachterijen van
ANB.
}
\seealso{
Other dataset: 
\code{\link{CLC18_BE}},
\code{\link{drg_example}}
}
\concept{dataset}
\keyword{datasets}
