% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check}
\usage{
check(x)
}
\arguments{
\item{x}{environment object}
}
\value{
1 = object exists in environment
0 = object doesn't exist in environment
}
\description{
Helper script to determine existence in environment panel
}
\author{
Sander Devisscher
}
