% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_osm_features.R
\name{collect_osm_features}
\alias{collect_osm_features}
\title{collect OpenStreetMaps features}
\usage{
collect_osm_features(proj_bbox)
}
\arguments{
\item{proj_bbox}{a bbox. The bounding box for the project/ study area for which
to extract osm features.}
}
\value{
a named list of 3 sf data frames:
osm_polygons, osm_lines, osm_points. Each sf data frame contains the
corresponding geometry types.
}
\description{
Extracts spatial features from the OpenStreetMaps server:
features that are extracted are re-classified into broad categories:
\itemize{
\item osm_polygons: urban, agriculture, open, forest, water
\item osm_lines: roads, waterways
\item osm_points: city names
}
}
\details{
A function to collect custom osm features for a project

dplyr and osmdata are automatically installed when missing from your system.
}
\examples{
\dontrun{

# extract the bounding box (WGS84) for the Project
proj_sf <- st_sfc(st_polygon(list(drg_example$spatial$coordinates[1,,])), crs = 4326)
proj_bbox <- st_bbox(proj_sf)
class(proj_bbox)
# extract selected OSM features
osm <- collect_osm_features(proj_bbox)

# plot
## polygons
(p1 <- ggplot(osm$osm_polygons \%>\% filter(!is.na(landuse)) \%>\% arrange(landuse)) +
   geom_sf(aes(fill = landuse), col = NA) +
   scale_fill_manual(values = unique(arrange(osm$osm_polygons, landuse)$osm_fill)) +
   theme_void() + theme(legend.position = "right"))

## lines
(p2 <- ggplot(osm$osm_lines) +
   geom_sf(aes(col = landuse)) +
   scale_color_manual(values = c("grey50", "#0092da")) +
   theme_void() + theme(legend.position = "right"))

## points
(p3 <- ggplot(osm$osm_points) + geom_sf_label(aes(label = name)))

## combine features
p1 + geom_sf(data = osm$osm_lines, aes(col = landuse)) +
 geom_sf_label(data = osm$osm_points, aes(label = name)) +
 scale_color_manual(values = c("grey20", "#0092da")) +
 coord_sf(xlim = proj_bbox[c("xmin", "xmax")],
          ylim = proj_bbox[c("ymin", "ymax")])
}
}
\author{
Martijn Bollen
}
