% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRS_extracter.R
\name{CRS_extracter}
\alias{CRS_extracter}
\title{CRS_extracter}
\usage{
CRS_extracter(CRS, EPSG = TRUE)
}
\arguments{
\item{CRS}{A character string with the name of the CRS}

\item{EPSG}{A logical indicating whether the output should be in EPSG format}
}
\value{
A CRS object
}
\description{
Extracts a coordinate reference system (CRS) from a library of commonly used CRS codes
}
\examples{
\dontrun{
crs_wgs <- CRS_extracter("WGS", EPSG = FALSE)
crs_bel72 <- CRS_extracter("BEL72", EPSG = FALSE)

epsg_wgs <- CRS_extracter("WGS", EPSG = TRUE)
}

}
\author{
Sander Devisscher
}
