% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_to_decimal.R
\name{dms_column_to_decimal}
\alias{dms_column_to_decimal}
\title{dms column_to_decimal}
\usage{
dms_column_to_decimal(df, coord_col)
}
\arguments{
\item{df}{Data frame containing the DMS columns.}

\item{coord_col}{Name of the column with coordinates in degrees.}
}
\value{
Data frame with a new column containing decimal degrees.
}
\description{
Convert a data frame column with degrees, minutes as well as seconds to decimal degrees.
}
\examples{
\dontrun{
# Example data frame
df <- data.frame(
 lat_Y = c("50\u00B045'38.7\"N", "50\u00B046'04.9\"N", "50\u00B045'19.7\"N",
 "50\u00B045'18.0\"N", "50\u00B046'24.4\"N", "50\u00B045'29.2\"N"),
 long_X = c("3\u00B010'26.0\"E", "3\u00B010'28.0\"E", "3\u00B010'25.0\"E",
 "3\u00B010'24.0\"E", "3\u00B010'30.0\"E", "3\u00B010'27.0\"E"))

# Convert the lat_Y and long_X columns to decimal degrees
df <- dms_column_to_decimal(df, c("lat_Y", "long_X"))
# View the updated data frame
print(df)
}
}
\seealso{
Other spatial: 
\code{\link{CRS_extracter}()},
\code{\link{add_habitats}()},
\code{\link{apply_grtsdb}()},
\code{\link{calculate_polygon_centroid}()},
\code{\link{collect_osm_features}()},
\code{\link{dms_to_decimal}()}
}
\author{
Sander Devisscher
}
\concept{spatial}
