% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_to_decimal.R
\name{dms_to_decimal}
\alias{dms_to_decimal}
\title{dms_to_decimal}
\usage{
dms_to_decimal(deg, min, sec, direction = NULL)
}
\arguments{
\item{deg}{Degrees (numeric or character).}

\item{min}{Minutes (numeric or character).}

\item{sec}{Seconds (numeric or character).}

\item{direction}{Direction (character, optional, e.g., 'N', 'S', 'E', 'W').}
}
\value{
Decimal degrees (numeric).
}
\description{
Convert degrees, minutes, and seconds to decimal degrees.
}
\examples{
\dontrun{
# Convert 30\u00B0 15'50"N or 30 degrees, 15 minutes, and 50 seconds north to decimal degrees
dms_to_decimal(30, 15, 50, 'N')
# Returns 30.26389
}
}
\seealso{
Other spatial: 
\code{\link{CRS_extracter}()},
\code{\link{apply_grtsdb}()},
\code{\link{calculate_polygon_centroid}()},
\code{\link{collect_osm_features}()},
\code{\link{dms_column_to_decimal}()}
}
\author{
Sander Devisscher
}
\concept{spatial}
