% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_seq_media.R
\name{download_seq_media}
\alias{download_seq_media}
\title{Download sequence media}
\usage{
download_seq_media(dataset, seqID, favorite = FALSE, outputfolder = NULL)
}
\arguments{
\item{dataset}{character string, path to the folder where a camptraptor datapackage has been unzipped.}

\item{seqID}{character string, ID of the sequence to download media from}

\item{favorite}{boolean, do you only want the pretty pictures?}

\item{outputfolder}{character string, path where the function should download the media into}
}
\value{
Downloads the specified media files into the outputfolder
}
\description{
This function allows the user to download all media related to a Agouti -
sequence which matches the given parameters.
}
\details{
If you are getting an Authorization Error (#403), this probably means your Agouti project has Restrict Images on. This needs to be turned off.
}
\examples{
\dontrun{
drg <- fistools::drg_example

# Situation 1: download whole sequence
download_seq_media(dataset = drg,
                    seqID = "f4c049d2-d42f-4cd3-a951-fd485ed0279a")

# Situation 2: download only favorited species media within sequence
download_seq_media(dataset = drg,
                    seqID = "f4c049d2-d42f-4cd3-a951-fd485ed0279a",
                    favorite = TRUE)
}

}
\seealso{
Other download:
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_gdrive_if_missing]{download_gdrive_if_missing()}},
\code{\link[=get_last_modified_zip]{get_last_modified_zip()}},
\code{\link[=read_dmogg]{read_dmogg()}},
\code{\link[=unzip_last_modified_zip]{unzip_last_modified_zip()}}
}
\author{
Lynn Pallemaerts

Emma Cartuyvels

Sander Devisscher

Soria Delva
}
\concept{download}
