% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_example.R
\docType{data}
\name{drg_example}
\alias{drg_example}
\title{drg_example}
\format{
datapackage
\describe{
\item{deployments}{Lijst van geselecteerde deployments}
\item{observations}{Observaties van de geselecteerde deployments}
\item{media}{Lijst van media-urls van de geselecteerde deployments}
}
}
\source{
\url{https://www.agouti.eu}
}
\usage{
drg_example
}
\description{
Subset of Drongengoed Agouti export for testing of functions
}
\seealso{
Other dataset:
\code{\link{CLC18_BE}},
\code{\link{boswachterijen}},
\code{\link{surveyMonkey_q6}}

Other agouti:
\code{\link[=agouti_imager]{agouti_imager()}},
\code{\link[=agouti_validate_ai]{agouti_validate_ai()}},
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link[=rename_ct_files]{rename_ct_files()}}
}
\concept{agouti}
\concept{dataset}
\keyword{datasets}
