% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_example.R
\docType{data}
\name{drg_example}
\alias{drg_example}
\title{drg_example}
\format{
datapackage
\describe{
\item{deployments}{Lijst van geselecteerde deployments}
\item{observations}{Observaties van de geselecteerde deployments}
\item{media}{Lijst van media-urls van de geselecteerde deployments}
}
}
\source{
\url{https://www.agouti.eu}
}
\usage{
drg_example
}
\description{
Subset of Drongengoed Agouti export for testing of functions
}
\seealso{
Other dataset:
\code{\link{CLC18_BE}},
\code{\link{boswachterijen}},
\code{\link{surveyMonkey_q6}}
}
\concept{dataset}
\keyword{datasets}
