% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wekeo_data.R
\name{get_wekeo_data}
\alias{get_wekeo_data}
\title{Get Wekeo data using Earthkit}
\usage{
get_wekeo_data(
  dataset_id,
  api_request,
  bbox = NULL,
  productType,
  resolution,
  startdate,
  enddate,
  itemsPerPage = 200,
  startIndex = 0
)
}
\arguments{
\item{dataset_id}{The dataset identifier.}

\item{api_request}{The API request in JSON format.}

\item{bbox}{A bounding box for the data request (optional).}

\item{productType}{The type of product to request (optional).}

\item{resolution}{The resolution of the data (optional).}

\item{startdate}{The start date for the data request (optional).}

\item{enddate}{The end date for the data request (optional).}

\item{itemsPerPage}{The number of items per page (default is 200).}

\item{startIndex}{The starting index for pagination (default is 0).}
}
\value{
A terra SpatRaster object containing the requested data.
}
\description{
Get Wekeo data using Earthkit
}
\details{
This function retrieves data from the \href{https://wekeo.copernicus.eu/}{Wekeo service} using the Earthkit library.
It constructs an API request based on the provided parameters and downloads the data.
This function is a wrapper around the \code{earthkit_download} from python function.
To get this function to work, you need to have the \code{earthkit} python package installed
which depends on the \verb{Microsoft C++ Build Tools}.
This toolkit is required to compile the \code{earthkit} package.
To install the toolkit please place ict helpdesk call at \code{ict.helpdesk@inbo.be}

When no API request is provided the function constructs a custom API request.
When the API request is provided, it will be used directly, ignoring the other parameters (except dataset_id).

This function uses .hdarc file to store the HDA credentials.
If the file does not exist, the user will be prompted to enter their HDA username and password.
These are your copernicus login & password.
}
\examples{
\dontrun{

# Set up the virtual environment and install required python packages
py_env <- "myenv" # Change this to your desired virtual environment name
# 0. Install python and create the virtualenv if needed
if(!reticulate::py_available()) {
 reticulate::install_python()
}
if (!py_env \%in\% reticulate::virtualenv_list()) {
 reticulate::virtualenv_create(py_env)
}

# 1. Activate the environment for the session
reticulate::use_virtualenv(py_env, required = TRUE)

# 2. Install Python packages if not already installed
if (!reticulate::py_module_available("hda")) {
  message("Installing Python module 'hda'...")
  reticulate::virtualenv_install(py_env,"hda")
}
if (!reticulate::py_module_available("earthkit")) {
  message("Installing Python module 'earthkit'...")
  reticulate::virtualenv_install(py_env,"earthkit")
}

# Example usage with a custom API request
api_request <- '{
"dataset_id": "EO:CLMS:DAT:CLMS_GLOBAL_NDVI_300M_V1_10DAILY_NETCDF",
"resolution": "300",
"bbox": [
  1.6063443461063671,
  48.05229722213296,
  8.35299059975311,
  51.7736550957488
],
"startdate": "2021-01-01T00:00:00.000Z",
"enddate": "2021-01-01T23:59:59.999Z",
"itemsPerPage": 200,
"startIndex": 0
}'

data <- get_wekeo_data(
  dataset_id = "EO:CLMS:DAT:CLMS_GLOBAL_NDVI_300M_V1_10DAILY_NETCDF",
  api_request = api_request
) |> terra::rast()
}
}
\seealso{
Other download: 
\code{\link{download_dep_media}()},
\code{\link{download_gdrive_if_missing}()},
\code{\link{download_seq_media}()}
}
\author{
Martijn Bollen & Sander Devisscher

This function retrieves data from the Wekeo service using the Earthkit library.
}
\concept{download}
