% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_selecter.R
\name{label_selecter}
\alias{label_selecter}
\title{label_selecter}
\usage{
label_selecter(
  label,
  update = FALSE,
  label_type,
  jaar,
  soort,
  bo_dir = "~/Github/backoffice-wild-analyse/",
  debug = FALSE,
  email = Sys.getenv("email")
)
}
\arguments{
\item{label}{een character (lijst) met labelnummer(s) die dienen onderzocht te worden. Dit kan in 3 vormen (volgnummer, met streepjes of zonder streepjes) of een combinatie van deze vormen aangeleverd worden}

\item{update}{een boolean die aangeeft of ook de nog niet wegeschreven dwh - bestanden moeten worden gecontroleerd.}

\item{label_type}{een character (lijst) met labeltypes die dienen onderzocht te worden.}

\item{jaar}{een numerieke (lijst) van jaren die dienen onderzocht te worden.}

\item{soort}{een character van de soort die onderzocht dient te worden.}

\item{bo_dir}{een character met de directory waar de backoffice-wild-analyse repository staat.}

\item{debug}{een boolean die aangeeft of de debug modus moet worden aangezet.}

\item{email}{een character met het email adres van de gebruiker. Wordt gebruikt voor authenticatie bij het updaten van de google drive bestanden. Standaard wordt het email adres uit de system variables gehaald, indien deze niet bestaat zal er een popup verschijnen waarin je je email adres kan ingeven.}
}
\value{
Een dataframe met de volgende kolommen:
\itemize{
\item INPUTLABEL: de input label
\item LABELTYPE: de labeltype(s) die onderzocht worden
\item JAAR: het jaar waarin de labels onderzocht worden
\item AM_OLD: een boolean die aangeeft of de label(s) in AfschotMelding voorkomen \strong{voor} de update van DWH_Connect
\item AM_OLD_LABEL: de label(s) die in AfschotMelding voorkomen \strong{voor} de update van DWH_Connect
\item TL_OLD: een boolean die aangeeft of de label(s) in ToegekendeLabels voorkomen \strong{voor} de update van DWH_Connect
\item TL_OLD_LABEL: de label(s) die in ToegekendeLabels voorkomen \strong{voor} de update van DWH_Connect
\item TL_CLEANED: een boolean die aangeeft of de label(s) in Toekenningen_Cleaned voorkomen
\item TL_CLEANED_LABEL: de label(s) die in Toekenningen_Cleaned voorkomen
\item DMOG: een boolean die aangeeft of de label(s) in Dieren_met_onderkaakgegevens voorkomen
\item DMOG_LABEL: de label(s) die in Dieren_met_onderkaakgegevens voorkomen
\item DMOG_GEO: een boolean die aangeeft of de label(s) in Dieren_met_onderkaakgegevens_Georef voorkomen
\item DMOG_GEO_LABEL: de label(s) die in Dieren_met_onderkaakgegevens_Georef voorkomen
\emph{Als \code{update = TRUE} worden de volgende kolommen toegevoegd:}
\item AM_NEW: een boolean die aangeeft of de label(s) in AfschotMelding voorkomen \strong{na} de update van DWH_Connect
\item AM_NEW_LABEL: de label(s) die in AfschotMelding voorkomen \strong{na} de update van DWH_Connect
\item TL_NEW: een boolean die aangeeft of de label(s) in ToegekendeLabels voorkomen \strong{na} de update van DWH_Connect
\item TL_NEW_LABEL: de label(s) die in ToegekendeLabels voorkomen \strong{na} de update van DWH_Connect
}
}
\description{
Deze functie onderzoekt of de labels bestaan in de datasets AfschotMelding (AM), ToegekendeLabels (TL), Toekenningen_Cleaned (TL_Cleaned), Dieren_met_onderkaakgegevens (DMOG), Dieren_met_onderkaakgegevens_Georef (DMOGG).
}
\details{
De parameter \code{label_type}, \code{jaar} en \code{soort} zijn enkel relevant als één van
de labels de vorm 'volgnummer' heeft. Wanneer deze parameter niet gespecifieerd
worden zal een default waarde voor het jaar (2013 t.e.m. max(AfschotMelding$Jaartal))
en label_type (c("REEGEIT", "REEKITS", "REEBOK", "WILD ZWIJN", "DAMHERT", "EDELHERT"))
gebruikt worden. Wanneer soort gespecifieerd is zal de lijst van labeltypes
beperkt worden tot deze die op de soort betrekking hebben. Voor ree bvb wordt dit reekits, reegeit en reebok.

De parameters \code{label}, \code{label_type}, \code{jaar} en \code{soort} kunnen als lijst aangeleverd worden.

De parameters \code{label_type}, \code{jaar} en \code{soort} zijn niet hoofdlettergevoelig.

\code{bo_dir} is de directory waar de backoffice-wild-analyse repository staat.
De functie checkt namelijk of de labels voorkomen in de lokale versie van de backoffice-wild-analyse repository.
Hiervoor is het belangrijk dat de backoffice-wild-analyse repository lokaal aanwezig is en de laatste versie gepulled is.

\code{update} is een boolean die aangeeft of de nog niet wegeschreven dwh - bestanden moeten worden gecontroleerd.
om dit te kunnen lopen is een verbinding met de DWH nodig. Dit is enkel mogelijk als je met de VPN van het INBO verbonden bent.
Of als je aanwezig bent op een vestiging van de Vlaamse Overheid (VAC).
}
\examples{
\dontrun{
#enkel label:
 label <- c(1234, "ANB2016REEGEIT001234", "ANB-2016-REEGEIT001234")
 output <- label_selecter(label)

#label & labeltype
 label <- c(1234, "ANB2016REEGEIT001234", "ANB-2016-REEGEIT001234")
 labeltype <- c("reegeit", "REEBOK")
 output <- label_selecter(label, label_type = labeltype)

#label & jaar & soort
 label <- c(1234, "ANB2016REEGEIT001234", "ANB-2016-REEGEIT001234")
 soort <- "ree"
 jaar <- c(2018, 2019)
 output <- label_selecter(label, jaar = jaar , soort = soort)
}
}
\seealso{
Other other:
\code{\link[=retry_function]{retry_function()}}
}
\author{
Sander Devisscher
}
\concept{other}
