% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renormalize_line_endings.R
\name{normalize_line_endings}
\alias{normalize_line_endings}
\title{Normalize line endings according to \code{.gitattributes}}
\usage{
normalize_line_endings()
}
\value{
Bestanden worden, indien relevant, aangepast, en de gebruiker krijgt de optie om deze veranderingen meteen te committen.
}
\description{
Deze functie herwerkt line endings (LF of CRLF) van alle bestanden in de huidige directory
en subdirectories, op basis van de regels in het \code{.gitattributes}-bestand. Dit zorgt voor consistentie tussen
de online repository en de bestanden op de lokale computer
}
\details{
De functie
\itemize{
\item Controleert eerst of er een \code{.gitattributes}-bestand aanwezig is.
\item Voert vervolgens de renormalisatie van alle bestanden uit met Git.
\item Controleert of bestanden aangepast zijn en vraagt of deze mogen gecommit worden.
\item Geeft statusmeldingen weer over de voortgang en het resultaat.
}
}
\examples{
\dontrun{
# Run renormalization and (optionally) commit changes
normalize_line_endings()
}
}
\author{
Soria Delva
}
