% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dmogg.R
\name{read_dmogg}
\alias{read_dmogg}
\title{Read "dieren met onderkaakgegevens georef"}
\usage{
read_dmogg(type = "short", email = Sys.getenv("email"))
}
\arguments{
\item{type}{"short" or "long" indicating which version you want to import.
The default is "short". See details for more info.}

\item{email}{email used to authenticate the google drive}
}
\value{
a dataframe containing the most recent DMOGG data.
}
\description{
This function reads the most up to date version of the
dieren met onderkaakgegevens georef (DMOGG) file from the backoffice wild
analyse.
}
\details{
This function can import two versions of the DMOGG dataset:
\itemize{
\item The \emph{"Short"} version: Contains data collected via the e-loket and, later,
the ANB's WIZ-app, enriched with data from the autopsiedb. Its temporal
range spans from 2014 to the present. This dataset is more comprehensive
and includes spatial data for WBEs and FBZs. It also contains data on
causes of death other than culling (mainly "valwild") and exotic species
(such as raccoon, muntjac, and sika deer).
\item The \emph{"Long"} version: Combines older data collected by INBO via e-loket
predecessors (e.g., zwijntje) with the subset of the "Short" version
that specifically concerns culling (\code{doodsoorzaak == "afschot"}) and a
select group of species (e.g., roe deer, wild boar, red deer, and fallow
deer). This dataset is less comprehensive because the early data
(pre-2014) lacks almost all spatial information.
}

This function uses \code{fistools::download_gdrive_if_missing()} under the hood to
download the file from the google drive.
}
\examples{
\dontrun{
# read the most recent short DMOGG data
dmogg <- read_dmogg()
}

}
\seealso{
Other download:
\code{\link[=download_dep_media]{download_dep_media()}},
\code{\link[=download_gdrive_if_missing]{download_gdrive_if_missing()}},
\code{\link[=download_seq_media]{download_seq_media()}},
\code{\link[=get_last_modified_zip]{get_last_modified_zip()}},
\code{\link[=unzip_last_modified_zip]{unzip_last_modified_zip()}}
}
\author{
Sander Devisscher
}
\concept{download}
