% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_ct_files.R
\name{rename_ct_files}
\alias{rename_ct_files}
\title{Rename camera trap files for upload in Agouti}
\usage{
rename_ct_files(foldername, extensions = c("jpg", "png", "jpeg", "bmp"))
}
\arguments{
\item{foldername}{DCIM folder to be treated}

\item{extensions}{file extensions to be rename. Default is jpg, jpeg, bmp and png}
}
\value{
renamed picture files in the original folder
}
\description{
Function that rename camera trap files (i.e. photos) by
appending folder name (e.g. 1000RECX) to file name (e.g. IMG0001) to create
all unique file names across a deployment. Needed for deployments with more
than 9999 pictures, so they can be uploaded in Agouti. This is a copy of the
old \code{rename_file_names()} in the old \code{fis-projecten} (archived).
}
\examples{
\dontrun{
fistools::download_gdrive_if_missing(gfileID = "1-84hbKouLKGwnLgBSRaQO1BnfogoFZWz",
                                     destfile = tempfile(fileext = ".zip"),
                                     email = Sys.getenv("email"),
                                     update_always = TRUE)
unzip(tempfile(fileext = ".zip"),
      exdir = "./data/test_rename_ct_files_unzipped")
foldername <- "./data/test_rename_ct_files_unzipped"
rename_file_names(foldername)
unlink("./data/test_rename_ct_files_unzipped",
       recursive = TRUE)
}
#'@export
}
\author{
: Lynn Pallemaerts and Someone Before Her
}
