% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_function.R
\name{retry_function}
\alias{retry_function}
\title{Retry a function multiple times}
\usage{
retry_function(expr, max_attempts = 3)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{max_attempts}{The maximum number of attempts to make.}
}
\value{
The result of the expression if successful.
}
\description{
This function evaluates an expression multiple times until it succeeds or the maximum number of attempts is reached.
}
\examples{
\dontrun{
# This example will fail the first two times, but succeed on the third attempt.
retry_function({
 if (runif(1) < 0.5) {  # Randomly fail
 stop("Failed")
 }
 print("Success")
 })
}
}
\seealso{
Other other: 
\code{\link{label_selecter}()}
}
\author{
Sander Devisscher
}
\concept{other}
