% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunsetter.R
\name{sunsetter}
\alias{sunsetter}
\title{Calculate the sunrise and sunset times for a given date and location}
\usage{
sunsetter(
  StartDate = Sys.Date(),
  EndDate = Sys.Date(),
  lat = 50.866572,
  lng = 4.350309
)
}
\arguments{
\item{StartDate}{Date in \%y-\%m-\%d format indicating startdate of dataframe. Defaults to today}

\item{EndDate}{Date in \%y-\%m-\%d format indicating enddate of dataframe. Defaults to today}

\item{lat}{Numeric indicating the latitude. Defaults to Herman Teirlinck building in Brussels}

\item{lng}{Numeric indicating the longitude. Defaults to Herman Teirlinck building in Brussels}
}
\value{
dataframe containing the dates between the startdate and enddate, the corresponding sunrise time and sunset time.
}
\description{
Calculate the sunrise and sunset times for a given date and location
}
\details{
This function uses the sunrise-sunset API to calculate the sunrise and sunset times for a given date and location.
}
\examples{
\dontrun{
# Example of how to use the sunsetter function
sunsetter(StartDate = "2021-01-01", EndDate = "2021-01-10", lat = 50.866572, lng = 4.350309)
}

}
\author{
Sander Devisscher
}
\concept{temporal}
