% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunsetter.R
\name{sunsetter2}
\alias{sunsetter2}
\title{sunsetter2}
\usage{
sunsetter2(df)
}
\arguments{
\item{df}{A dataframe containing the dates, latitudes and longitudes}

\item{dates}{A vector of dates contained in the dataframe}

\item{lat}{A vector of latitudes contained in the dataframe}

\item{lng}{A vector of longitudes contained in the dataframe}
}
\value{
dataframe containing the dates, latitudes, longitudes, sunrise time and sunset time.
}
\description{
Calculate the sunrise and sunset times for a given set of date and location combinations
}
\details{
This function uses the sunrise-sunset API to calculate the sunrise and sunset
times for a given set of date and location combinations.
}
\examples{
\dontrun{
# create a dataframe with dates, latitudes and longitudes
df <- data.frame(dates = c("2021-01-01", "2021-01-01", "2020-12-25"),
                 location = c("Brussels", "Amsterdam", "Brussels"),
                 lat = c(50.866572, 52.367573, 50.866572),
                 lng = c(4.350309, 4.904138, 4.350309),
                 remarks = c("New Year's Day", "New Year's Day", "Christmas Day"))

# calculate the sunrise and sunset times for the dataframe
sunsets <- sunsetter2(df)

# add the sunrise and sunset times to the dataframe
df <- dplyr::bind_cols(df, sunsets \%>\% dplyr::select(-dates))
df
}

}
\seealso{
Other temporal: 
\code{\link{sunsetter}()}
}
\author{
Sander Devisscher
}
\concept{temporal}
